chore: move worker profile to separate module
