+++ packages/apps/human-app/frontend/src/modules/worker/profile/components/kyc-verification-control.tsx
@@ -1,8 +1,7 @@
 import { useTranslation } from 'react-i18next';
-import { DoneLabel } from '@/modules/worker/components/profile/components/status-labels/done-label';
-import { ErrorLabel } from '@/modules/worker/components/profile/components/status-labels/error-label';
-import { StartKycBtn } from '@/modules/worker/components/profile/components/buttons/start-kyc-btn';
-import type { WorkerProfileStatus } from '@/modules/worker/components/profile/types/profile-types';
+import { type WorkerProfileStatus } from '../types/profile-types';
+import { StartKycBtn } from './buttons';
+import { DoneLabel, ErrorLabel } from './status-labels';
 
 interface KycStatusProps {
   status: Pick<
@@ -11,7 +10,7 @@ interface KycStatusProps {
   >;
 }
 
-export function KycVerificationControl({ status }: KycStatusProps) {
+export function KycVerificationControl({ status }: Readonly<KycStatusProps>) {
   const { t } = useTranslation();
   const kycToComplete = !(status.kycApproved || status.kycDeclined);
