+++ packages/apps/human-app/frontend/src/modules/worker/profile/hooks/use-worker-wallet-registration.ts
