+++ packages/apps/human-app/frontend/src/modules/worker/profile/index.ts
@@ -1,0 +1,1 @@
+export * from './views/profile.page';
